API

NamedTrajectory methods

Base.:== — Method

:(==)(traj1::NamedTrajectory, traj2::NamedTrajectory)

Check if trajectories are equal w.r.t. using Base.:(==)

Base.copy — Method

copy(::NamedTrajectory)

Returns a copy of the trajectory.

Base.getindex — Method

getindex(traj, ts::AbstractVector{Int})::Vector{KnotPoint}

Returns the knot points at times ts.

Base.getindex — Method

getindex(traj, t::Int)::KnotPoint

Returns the knot point at time t.

Base.getindex — Method

getindex(traj, symb::Symbol)

Dispatches indexing of trajectories as either accessing a component or a property via getproperty.

Base.getproperty — Method

getproperty(traj, symb::Symbol)

Returns the component of the trajectory with name symb or the property of the trajectory with name symb.

Base.isequal — Method

isequal(traj1::NamedTrajectory, traj2::NamedTrajectory)

Check if trajectories are equal w.r.t. data using Base.isequal

Base.lastindex — Method

lastindex(traj::NamedTrajectory)

Returns the final time index of the trajectory.

Base.length — Method

length(::NamedTrajectory)

Returns the length of all variables of the trajectory, including global data.

Base.merge — Method

merge(traj1::NamedTrajectory, traj2::NamedTrajectory)
merge(trajs::AbstractVector{<:NamedTrajectory})

Returns a new NamedTrajectory object by merging NamedTrajectory objects.

Merge names are used to specify which components to merge by index. If no merge names are provided, all components are merged and name collisions are not allowed. If merge names are provided, the names are merged using the data from the index provided in the merge names.

Joined NamedTrajectory objects must have the same timestep. If a free time trajectory is desired, setting the keyword argument free_time=true will construct the a component for the timestep. In this case, the timestep symbol must be provided.

Arguments

traj1::NamedTrajectory: The first NamedTrajectory object.

traj2::NamedTrajectory: The second NamedTrajectory object.

free_time::Bool=false: Whether to construct a free time problem.

timestep_name::Symbol=:Δt: The timestep symbol to use for free time problems.

merge_names::Union{Nothing, NamedTuple{<:Any, <:Tuple{Vararg{Int}}}}=nothing: The names to merge by index.

Base.setproperty! — Method

setproperty!(traj, name::Symbol, val::Any)

Dispatches setting properties of trajectories as either setting a component or a property via setfield! or update!.

Base.size — Method

size(traj::NamedTrajectory) = (dim = traj.dim, T = traj.T)

Returns the size of the trajectory (dim, T), excluding global data.

TODO: Should global data be in size?

Base.vec — Method

vec(::NamedTrajectory)

Returns all variables of the trajectory as a vector, Z⃗.

NamedTrajectories.MethodsNamedTrajectory.add_component! — Method

add_component!(traj, name::Symbol, data::AbstractVecOrMat; type={:state, :control})

Add a component to the trajectory.

This function resizes the trajectory, so global components and components must be adjusted.

NamedTrajectories.MethodsNamedTrajectory.get_component_names — Method

get_component_names(traj::NamedTrajectory, comps::AbstractVector{<:Int})

Returns the name of the component with the given indices. If only one component is found, the name is returned as a single symbol. Else, the names are returned as a vector of symbols.

The filter requires that the components are a complete subset of the given indices, so that a partial match is excluded from the returned names.

NamedTrajectories.MethodsNamedTrajectory.get_components — Method

get_components(::NamedTrajectory)

Returns a NamedTuple containing the names and corresponding data matrices of the trajectory.

NamedTrajectories.MethodsNamedTrajectory.get_duration — Method

get_duration(::NamedTrajectory)

Returns the duration of a trajectory.

NamedTrajectories.MethodsNamedTrajectory.get_times — Method

get_times(traj)::Vector{Float64}

Returns the times of a trajectory as a vector.

NamedTrajectories.MethodsNamedTrajectory.get_timesteps — Method

get_timesteps(::NamedTrajectory)

Returns the timesteps of a trajectory as a vector.

NamedTrajectories.MethodsNamedTrajectory.merge_outer — Method

merge_outer(objs::AbstractVector{<:Any})

Merge objects. An error is reported if a key collision is detected.

NamedTrajectories.MethodsNamedTrajectory.remove_component — Method

remove_component(traj, name::Symbol)

Remove a component from the trajectory.

NamedTrajectories.MethodsNamedTrajectory.remove_components — Method

remove_components(traj, names::Vector{Symbol})

Remove a set of components from the trajectory.

NamedTrajectories.MethodsNamedTrajectory.update! — Method

update!(traj, datavec::AbstractVector{Float64})

Update the trajectory with a new datavec.

NamedTrajectories.MethodsNamedTrajectory.update! — Method

update!(traj, name::Symbol, data::AbstractMatrix{Float64})

Update a component of the trajectory.

NamedTrajectories.MethodsNamedTrajectory.update_bound! — Function

update_bound!(traj, name::Symbol, data::Real)
update_bound!(traj, name::Symbol, data::AbstractVector{<:Real})
update_bound!(traj, name::Symbol, data::Tuple{R, R} where R <: Real)

Update the bound of a component of the trajectory.

Random trajectory methods

Base.rand — Method

rand(
    ::Type{NamedTrajectory},
    T::Int;
    timestep_value::Float64=1.0,
    timestep_name::Symbol=:Δt,
    free_time::Bool=false,
    timestep::Union{Float64,Symbol}=free_time ? timestep_name : timestep_value,
    state_dim::Int=3,
    control_dim::Int=2
)

Create a random NamedTrajectory with T time steps, a state variable x of dimension state_dim, and a control variable u of dimension control_dim. If free_time is true, the time step is a symbol timestep_name and the time step value is timestep_value. Otherwise, the time step is a number timestep_value.